move file from components to component
